Extract logic from bookings_route to service
